src/modules/ingestion/ingestion.service.ts
import { Injectable } from '@nestjs/common';
import { User } from '../users/entities/user.entity';

@Injectable()
export class IngestionService {
  // Simulate an ingestion trigger
  async triggerIngestion(fileId: number, user: User) {
    return {
      message: `Ingestion triggered successfully for file ${fileId}`,
      userId: user.id,
    };
  }

  // Get the status of an ingestion job
  async getIngestionStatus(ingestionId: string, user: User) {
    // Simulate status check logic, or pull from a database if needed
    const status = 'In Progress'; // This would be dynamic in a real system
    return {
      status,
      ingestionId,
      userId: user.id,
    };
  }

  // Cancel an ongoing ingestion job
  async cancelIngestion(ingestionId: string, user: User) {
    // Here, you would implement logic to stop the ingestion job, such as aborting a background task
    return {
      message: `Ingestion canceled for ingestionId: ${ingestionId}`,
      ingestionId,
      userId: user.id,
    };
  }
}
